GCS/Video_Analytics_PilotV1/requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
bleach==1.5.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
ffmpeg-python==0.2.0
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
imageai==2.1.5
imutils==0.5.3
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2
matplotlib==3.1.3
numpy==1.17.4
oauthlib==3.1.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
parso==0.6.1
pickleshare==0.7.5
Pillow==6.2.1
progressbar==2.5
prompt-toolkit==3.0.3
protobuf==3.11.3
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==1.6.0
tensorflow==1.6.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.11.2
